Remove monitor mode
Delete all monitor mode UI code - the feature that displayed all tabs in a grid view.

Removed:
- internal/ui/center/monitor.go, monitor_test.go, monitor_tabs_test.go
- internal/ui/center/model_monitor.go, model_monitor_snapshots.go
- internal/app/app_monitor.go, app_monitor_grid.go
- internal/e2e/snapshot_test.go
- Monitor keybinding, toolbar button, help section
- All monitorMode state, snapshot caching, and related logic

Kept: tmux monitor-activity setting (unrelated feature)
Amp-Thread-ID: https://ampcode.com/threads/T-019c24a0-60c7-7516-9f83-ec3b4ce7e3df
