+++ .github/workflows/ci.yml
@@ -29,9 +29,6 @@ jobs:
       - name: Test
         run: go test ./...
 
-      - name: Harness (monitor)
-        run: go run ./cmd/amux-harness -mode monitor -frames 5 -warmup 1
-
       - name: Harness (center)
         run: go run ./cmd/amux-harness -mode center -frames 5 -warmup 1
