+++ internal/ui/dashboard/dashboard_render.go
@@ -212,7 +212,6 @@ func (m *Model) helpLines(contentWidth int) []string {
 	}
 	items = append(items, m.helpItem(focusKey, "focus (or ←↑↓→)"))
 	items = append(items,
-		m.helpItem("C-Spc m", "monitor"),
 		m.helpItem("C-Spc ?", "help"),
 		m.helpItem("C-Spc q", "quit"),
 	)
