+++ internal/ui/dashboard/model.go
@@ -43,7 +43,6 @@ type toolbarButtonKind int
 
 const (
 	toolbarHelp toolbarButtonKind = iota
-	toolbarMonitor
 	toolbarSettings
 )
